Improve borrow checks in protocol functions
The protocol functions can schedule a sync, which means that there is a
chance that a RefCell borrow fails.

All borrows are now reworked to their `try_borrow` variant, failing the
promise when the borrow fails. Most immutable borrows are very quick and
immediatly clone the storage, but the mutable borrows may take a long
time to complete which will fail the immutable borrow in the schedule
sync function. This is the more likely scenario and is kind of fine as
any other functions that currently hold a mutable borrow will sync the
store afterwards anyways.

This is not bulletproof, but it will no longer panic and break the
entire module as a result.